Fix async scheduling
sendAuth happens on the next event loop tick now that it's async, so
we need to wait on it to happen before resolving connect.

This issue manifests clearly when we start up and immediately issue a
sync request, such as create db or create user. The request fails
because we haven't transitioned yet to auth_sent.