Fix makefile variables for test target
When variable assignments appear as prerequisites of a make recipe,
the updated values are in effect when evaluating the body of the
recipe and the bodies of the dependent recipes. But in particular the
updated values are not in effect when evaluating the prerequisite
lists themselves. This causes the setting
`SKDB_BIN=sql/target/host/dev/skdb` as a prerequisite of `test` to not
override the global `SKDB_BIN=sql/target/host/release/skdb` when
determining if `build/skdb` is up-to-date. Instead use a recursive
invocation of make with updated variable values.