First implementation of js 'watch'
We use to force users to create virtual views before being able to
subscribe to changes. This approach was to heavy for the JS API. We
need to watch queries all the time and creating a virtual view (that
sticks forever) is not a reasonable approach.

So here comes reactive watches! The API:

watch(query, queryParams, (sqlRes) => void): { close: () => void }

So given an SQL query and query params, the watch will call the
closure passed as parameter every time a change affecting that query
occurs. The user can stop the stream by calling the 'close' method.

- renamed skdb.sql to skdb.exec in JS