Add context.update() to SKIP_reactive_query
Without this, the tick doesn't bump.

This means that each time reactiveQuery is called - without something
else causing a bump, say lots of queries on a react tree re-render -
the deps accumulate. Then when a node is removed from the tree it
wrongly deletes more deps than it should, causing the reactive queries
to stop reacting.