Filters with capacity in SKStore + incremental limits in SQL
This diff introduces a new construction called EagerFilter. An
EagerFilter works like a map except that it can only filter elements,
meaning, it will produce the same keys in the same order, but can
leave out some elements.

Of course, such a construction could be trivially encoded with a map,
so what's the point? Filters can be configured to have a certain
capacity. Concretely that means that if a filter has a capacity of 5,
it will only keep the first 5 elements, and it will do so very
efficiently (without having to scan the rest of the directory).

This feature is introduced to efficiently implement incremental limits
in SQL. We used to just compute the whole query, and then take out the
first N elements, and we used to ban the usage of "LIMIT" in virtual
views (because we were not able to update them incrementally
efficiently).

With this new SKStore construction, we can support LIMIT, even in an
incremental context.