lwwEdits speedup
Whenever we are inserting in a table with a primary key, we need to
lookup the elements that were there before to apply our lww
strategy. Unfortunately, the lookup was done by compiling/evaluating a
select query to perform that lookup, which is a bit more expensive
than necessary.

This diff changes that, by iterating on the index directly.