Fixed comparison function for SqlValues + fix lwwEdits
I hoisted out a check in SqlCsv.lwwEdits that was incorrect, because
outside of the loop the files are potentially empty.

The introduction of a shorter key in lwwEdits revealed a bug in the
comparison function SqlValue.sk. Nobody ever noticed because we only
used shorter keys in a context of an index, and in an index the order
that we compare the columns is the same as the order of the columns.