Do not --connect subscribe to establish a tail
This is a long hangover from when I didn't completely understand
subscribe semantics.

The connect ends up building the full change set, passing it to
notify, which because this is a tail that hasn't started yet, does
nothing.

For small tables this overhead is negligible. But for only 15k rows it
adds 50ms.