Fix: output latest tick checkpoint on init
This was a bug (I think it's actually fixed on the multi-peer branch)
that means clients don't receive the latest context tick on the
initial checkpoint. This is bad for efficiently resuming and probably
for the merge logic (looks like we've seen less than we have).

I caught it because we now don't resolve the mirror if the tick is 0,
so we were waiting the 10 secs for the cond var timeout on each mirror
when the table was empty.