New group model for skdb
It used to be that one would just add rows into skdb_group_permissions
to define a group. This model was simple and powerful, but the main
down side was that all permissions had to be created by the root user.

The reason why this is unfortunate is that there are many use cases
where a user would like to be able to create a group, therefore,
having to call a special endpoint every time greatly reduces the
appeal of skdb.

Of course, one could just define a field called 'skdb_access' on the
skdb_permissions and call it a day. The problem with this approach is
that it would not prevent anyone from modifying the group
permissions. One could still write into a group, just with a different
skdb_access.

So that's just not enough. The solution, implemented in this diff is
the following:

Create an entry in a table called skdb_groups:
 ('THE UUID OF THE GROUP',
  'THE AUTHOR' (can be null if root),
  'THE NAME OF THE ADMIN GROUP' (can be a userUUID),
  'THE ACCESS OF THIS ENTRY' (can user see this entry?)
 )

Once this is done, you can add entries to skdb_group_permissions, but
there is a catch: the every entry trying to use an skdb_access
different from the one defined in skdb_groups will be ignore.

Let's take an example.

```SQL
INSERT INTO skdb_groups
  VALUES ('ExampleGroup', 'user1', 'user1', 'read-only');
```

We just create a new group called ExampleGroup, the admin is only
'user1', but everybody can see that 'user1' is the admin.

```SQL
INSERT INTO skdb_group_permissions
  VALUES ('ExampleGroup', 'user2', skdb_permissions('r'), 'user1');
```

This entry will be taken into account, because the skdb_access entry
matches the one defined in skdb_groups. As a result, 'user2' is now a
member of 'ExampleGroup' and it is given the right to read.

However, the following:

```SQL
INSERT INTO skdb_group_permissions
  VALUES ('ExampleGroup', 'user2', skdb_permissions('r'), 'user2');
```

won't work, the entry will be ignored, because skdb_access (in this
case user2) differs from what was origingally defined in skdb_groups.