Update produces a multi-table diff
This lays the groundwork for the diff producing side. Mostly it's
about enriching the sub data model to talk about subscriptions to
multiple tables. Each may have it's own filtering.

The first milestone for x-table replication will leave tail as is -
single table output. The next milestone will enrich tail but to do
this we need to improve the watermark data model and the protocol to
talk in terms of multi-table tailing.

We break diffing in this commit but will fix that in a moment.