Do not do blocking work on IO threads
The IO threads are a small pool of event-loop style workers. Blocking
work not only limits throughput but can deadlock if it requires an I/O
thread to complete - which the websocket send/close methods do.

This is a quick fix to move all the work on to a worker thread. We
just use a single thread to ensure ordering. Long term we should
either use a sharded pool or - as has been the plan for a long time
now - we should move the mux logic to coroutines and make it all
async.