Make write-csv acks work again
We now generate the appropriate table header so that the client can
consume. This does mean that the table that receives acks is now part
of the protocol.

Also had to adjust some of the mechanics now that write-csv output is
multi-line.

This work sets us up well for write-csv to return rejected data.