Change diff interface for multi-table
Now we can provide --since (for backwards compat.) and this will be
used for all tables. Or we take a json-based input on stdin so that we
can diff from a point in time for each table.

This was designed so that diff knows nothing about watermarks or the
translation to response tables. It is also forward looking for the
tail interface which needs to be more complex (handle filters).