Implement echo filtering
This is almost correct but we need to only look at changes since the
tick we're interested in. Otherwise we can get in to a byzantine
generals situation where we need to keep replicating an updated row to
learn what changes have been received.

We'll do this in a follow up diff as it will be intrusive to the api
so best kept contained.