Support encoding/decoding strings of < 27 chars to ksuids
Add support for ksuids with text encoding that is < 27 characters, by
trimming trailing '-' characters when converting to the text form, and
right-padding shorter strings with '-' when converting from the text
form. The '-' char is chosen since it is the encoding of the 0 byte.

This change is currently not useful, but comes from thinking about
forward compatibility of e.g. using ksuids for the groupID column of
skdb_groups. With this change, a string such as 'read-write' can be
used as a valid text encoding of a ksuid, which is equivalent to
`read-write-----------------` but more human-friendly.