Perf optimisation
I hate myself for this commit. This is definitely premature. And
dangerous. The bottleneck is the skdb process creation and the python
process isn't CPU bound. But a test with a more complex check could
really benefit here. I'm going for it. Time will tell how much I
regret this.