Include hash of params in id of CSelect
Currently compiling a `Select` to a `CSelect` reuses the `id` if the
`Select`. In the presence of `params` that involve expressions such as
`id`, `local_sequence_number`, or `datetime('now')` where
`compileExpr` is not functional, it is possible for two distinct
`CSelect`s to have the same `id`. This leads to incorrect reuse of
previously computed results.

This diff fixes this by adding a hash of the `CSelect` `params` field
to the `id`.

This is not done for virtual views. Their ids must be preserved by
compilation, and it is the user's responsibility to name them
uniquely.