[skdb] Return error from `eval()`.
We previously relied on writing to `stderr` to communicate errors to
the caller, which worked fine when invoking `skdb` as a sub-process or
through wasm (where we capture and parse stderr), but would have made
the situation difficult for proper FFI bindings (which will be useful
for eg. Python bindings).

Since an exception can be raised from various places, we have to rely
on a skstore global, which has the added benefit of guaranteeing the
`Error` object remains alive when the obstack is popped.

Note: Actual error reporting (along with proper line numbers) will be
reintroduced in a subsequence commit.