Build automated expectation checking
This is based on a checkpoint signal sent by the other client. Based
on the replication semantics, we should be guaranteed that the
database is a known consistent state for assertions.

These are currently commented out as we have races: x-table tail
replication isn't yet implemented.

We also use a timeout so as not to run queries in the watch callback.
This introduces a second race.

Once these issues are resolved, we can uncomment.