Fixed context leak
There was several places in the codebase where we were leaking
contexts. What that means concretely is that a context should have
been freed and it wasn't. This lead to situation where the database
was allocating a lot more than was necessary.

This diff fixes the leak by revisiting the whole logic. In the old
scheme, the context would have a reference count at zero when noone
was active. This is now changed to 1 (it was easier to reason
about). Also, there is now a new construction called withContext for
those who want to make sure that they are freeing the context
properly. The code in SqlTailer has been adapted to use that.

We are still leaking a context if we receive a signal at the wrong
time. But that should be fixed in a subsequent diff. Also, that does
not affect wasm, so probably not very urgent.