Enforce admin-ness for group permission changes; remove joined vview
Previously, we allowed anyone to write to skdb_group_permissions, then used the
skdb_group_permissions_joined virtual view to filter out any rows that didn't
exactly match the admin group.  This causes issues when we want to sent the rows
to users who may, for example, have read permissions but not admin permissions.

Instead, this special-cases admin checking at write time, allowing freedom to
set `skdb_access` on `skdb_group_permissions` rows as with any other table.
That also means we no longer need the separate joined virtual view, and the state
of the skdb_group_permissions table is the underlying truth, not the filtered
view on it.