Fix: applying chunks only accepts the first and last
The first because no watermark is seen yet.

The middle are ignored because they all carry the 0 checkpoint which
matches the recorded watermark from the first chunk.

The last because it carries a non-zero checkpoint.

This now ensures that all are applied by not recording zero
watermarks. This makes the use of zero as a checkpoint value
consistent for both write-csv and tail.