Fix: server sends down a reset while there is a local change
Imagine (0, foo) is sent to the server. Then (0, bar) is written
locally. This orders: (0, bar) < (0, foo). An unfortunate reset will
wipe this out. Of course another client may also have written (0, foo)
concurrently which should win and then this is communicated with a
reset. Impossible to tell which happened, but I judge the former is
more likely. This is a temporary measure while we remove resets.