Revert "Fix: server sends down a reset while there is a local change"
This reverts commit efb3b306b2a1236dbc2e92d3ad0c575fa0e0dde5.

We revert this and change the test expectation to show the problem we
introduce by reverting the fix.

It's preferrable to have this issue than to have racing resets on the
server not use the natural ordering. Racing resets on the server will
happen any time the server fails for some length of time. Without the
lww logic we get non-deterministic behaviour and the cluster does not
converge. The simulation doesn't currently demonstrate this as it only
has a single reset happening. But the soak catches this and fails
sometimes with dramatic data loss with this change in.