Wrap row loops with an obstack
Processing rows produces garbage O(n). This makes the garbage O(1).

To make this work we also move to unsafeGetDataIterWithoutTombs so
that we're not modifying a context and then blowing away the obstack
with any references. This is fine, we don't intend to track the
changes here and save ourselves the metadata.

This likely is introducing a perf hit now and should be done at a less
fine grain - i.e. every 2^x rows. Will measure and follow up.

This reduces memory in my test by 15%: 438Mb -> 372Mb.