Re-write getKeyChangesAfterIter to not use yield
The last commit causes chaos in wasm.

After a couple days of debugging, I narrowed down the issue to a bad
call_indirect. We end up derefing a null pointer and calling the wrong
function instead of SID__LE as we should.

I didn't manage to figure out where the memory was getting corrupted.
I timed out and just re-wrote the code that was invoking SID__LE and
on a hunch didn't use yield. This now seems to work exactly as
intended.