Make the JSON parser generic
We have a JSON parser defined in the standard library. The problem is
that it produces JSON objects of a specific type. The object's fields
are represented with a hashtable, among other things. This is not
ideal, there are cases where we want to parse JSON, but we don't want
the result to be stored this way. Of course, we could first parse the
JSON into this form, and then convert that into what we need, but that
would be wasteful.

Instead, this diff turns the parser into something more generic. It
parametrizes the parser with a class, so that one can define different
ways to construct the JSON object.