Support for table level functions + JSON syntax.
Sometimes, we will want to define table level operations in the form
of a virtual view. More concretely, views of the form:

CREATE VIRTUAL VIEW my_view_name AS my_skip_fun(arg1, ..., argN);

Typically, my_skip_fun will take as parameter table names and
literals.

This diff introduces the syntax for that.

The diff also introduces syntax for a new type called JSON and BLOB.
JSON is actually exposed to the user, while BLOB is not (yet).