Support for JSON in SKDB
This diff introduces a new type: JSON

A JSON type is an object that is represented internally with an array
of values, coupled with an array of field names. The field names are
interned, so if two objects share the same field names, the table will
be shared between the two objects.

JSON objects cannot be decomposed directly in SQL, instead, we
introduce a new primitive that operates at the table level: json_split.

This primitive can only be used in a virtual view (for now?). It takes
as parameter an sql table name, that table is expected to have at
least 2 columns, where the first one is expected to be an TEXT ID and
the second one either TEXT or JSON (containing a JSON object).

CREATE VIRTUAL VIEW paths AS json_split(my_table);

This virtual view creates something of the form:
ID|PATH|TYPE|INTEGER|FLOAT|TEXT

For example, given the object {"a": 3}, the split will produce:

'f7xB8m2VDA5JqzOW4tkSlX_7Lw7'|.a|'int'|3|NULL|NULL

The idea is then to let people do their queries on this form.