Change of the representation of indexes keys
We use to represent keys of an index in a very inefficient way: this
was mostly due to accommodate "INSERT OR REPLACE". That construction
requires meta-data to understand the nature of the conflict. In this
diff, we only reconstruct that meta-data when necessary, saving a lot
of space.

This is a 20% win in space for a table that defines an index.