Move definition of SKIP_String_eq from preamble to runtime
It seems that SKIP_String_eq is defined in the preamble only to be
able to have an i1 return type instead of i8. This is an optimization
that LLVM should be able to do itself, so this PR simplifies the code
by moving the definition from LLVM assembly in the prelude to C in the
runtime.