[skargo] Compute the default number of jobs
On linux or macos machines with coreutils, `nproc` returns the number
processors. This PR changes skargo to use this to compute the default
number of parallel jobs.

For potential future reference, if we want to support vanilla macos,
we could first test if `uname -s` is `Darwin` and then execute `sysctl
hw.ncpu | cut -d' ' -f2` instead.