[skstore] Fix MInfo creation
Add addToUpdate function to help debug.

MInfo is a structure used to saves some file dependencies data (generated keys, created subdirectories and reads that have taken place during a mapped file creation).
Until now, when only one key was created, the created directories and reads were systematically ignored, with the result that when the source file was deleted, the generated subdirectories were not deleted.
Fixing this problem was relatively straightforward: save the list of directories and their reads, when exists.
In doing so, it revealed issues occurring when deleting subdirectories, such as in the case of watch, when a watch was deleted, the new Minfo was empty and therefore removed from the structure conserving this information.
However, the set of dependencies meant that the file in question was updated again, and as the information that the MInfo had not been explicitly saved, the update retrieved the information relating to the fixed data, which caused the entire dependency chain to be re-looped, generating an infinite loop.
Once identified, fixing this problem was straightforward: always save the MInfo.