[skstore] Add a way to share sub dirs between files
The second consequence was that it was no longer possible to share existing SELECT at VIRTUAL VIEW level, the principle being that is when a view entry is created, it checks whether a table relative to the SELECT already exists, otherwise it creates it.
Up until now, there had been no problem, as in the case where sub-directory creation data was not saved, SELECT related directories were never deleted. However, when these are present, all entries in the virtual view are considered as owner of the SELECT related directories, so when the first entry is deleted, all the directories are deleted, which has no consequence in most cases, since they are regenerated when required. The only issues arises when the SELECT contains an aggregate function.
In that case, the directory performing the aggregate computation is also deleted, with the result that it is not updated.
Several solutions were considered to fix that issue, but they all had a heavy impact on the code.
It turns out that adding a reference counter for sub-directory owners solves the issue with very little impact on the code.
And to avoid saving useless data, the ability to identify directories authorizing multiple owners was added.