Support for JSON in skdb (#66)
* Make the JSON parser generic

We have a JSON parser defined in the standard library. The problem is
that it produces JSON objects of a specific type. The object's fields
are represented with a hashtable, among other things. This is not
ideal, there are cases where we want to parse JSON, but we don't want
the result to be stored this way. Of course, we could first parse the
JSON into this form, and then convert that into what we need, but that
would be wasteful.

Instead, this diff turns the parser into something more generic. It
parametrizes the parser with a class, so that one can define different
ways to construct the JSON object.

* Introducing skjson

skjson is a library (and a binary) to manipulate JSON. Among the
things that it can do:

To generate random JSON objects (--n is the number of objects needed):
skjson gen-json --n 100

To infer a schema:
cat my_json_objects | skjson infer-schema

To check against a schema:
cat my_json_objects | skjson check --schema '{"a":int}'

etc ...

This tool/library is needed to properly support JSON in skdb.

* Support for table level functions + JSON syntax.

Sometimes, we will want to define table level operations in the form
of a virtual view. More concretely, views of the form:

CREATE VIRTUAL VIEW my_view_name AS my_skip_fun(arg1, ..., argN);

Typically, my_skip_fun will take as parameter table names and
literals.

This diff introduces the syntax for that.

The diff also introduces syntax for a new type called JSON and BLOB.
JSON is actually exposed to the user, while BLOB is not (yet).

* Support for JSON in SKDB

This diff introduces a new type: JSON

A JSON type is an object that is represented internally with an array
of values, coupled with an array of field names. The field names are
interned, so if two objects share the same field names, the table will
be shared between the two objects.

JSON objects cannot be decomposed directly in SQL, instead, we
introduce a new primitive that operates at the table level: json_split.

This primitive can only be used in a virtual view (for now?). It takes
as parameter an sql table name, that table is expected to have at
least 2 columns, where the first one is expected to be an TEXT ID and
the second one either TEXT or JSON (containing a JSON object).

CREATE VIRTUAL VIEW paths AS json_split(my_table);

This virtual view creates something of the form:
ID|PATH|TYPE|INTEGER|FLOAT|TEXT

For example, given the object {"a": 3}, the split will produce:

'f7xB8m2VDA5JqzOW4tkSlX_7Lw7'|.a|'int'|3|NULL|NULL

The idea is then to let people do their queries on this form.

* Made sure that SKJSON errors always have "Error" in the message

+ renamed a few tests in sql/test_json.sh (because several had the
same name).

* introduced types for JSON and SCHEMA

* Implementation of json_extract

Example:

create virtual view foo as
  json_extract(tableName, colName, '{x<int>');

Will extract all the fields named x of type integer out of
talbeName.colName.

* Fixed duplicate wasm test error

* Added missing "module end;" in skjson

* Fixed support for num. int | float types in JSON should convert to float in SQL

* Proper utf8 support for stdin in js

We were reading the string character by character in UTF-16 but
interpreting it as UTF-8 on the wasm end.

* Revisit the function decoding utf8 using native primitives