Fix UTF8 in JS getChar (#111)
The way getChar was implemented in JS was not matching the native
implementation. The native version return a codepoint, the JS version
was supposed to do the same (using a construction called
.codePointAt), but it was not working on some utf8 encoded examples
(for reasons that I don't understand).

This diff mimics the logic that we have natively. It would be probably
worth revisiting at some point, because it would be nicer to use JS
primitives to do the work, but it's not urgent. Let's fix our JS
first.

The diff also includes tests on utf8 strings to make sure we don't
regress.