Represent stdin as a utf8 byte array and use TextEncoder
stdin is a byte stream and in my opinion should assume utf8. This
seems to be the assumption we have everywhere else. So I changed the
type to reflect this and then can use the built in TextEncoder to
convert strings.

Now this is JS so all bets are off, but I would hope this is likely
more correct and more efficient than anything we've hand rolled. But
it should definitely produce (slightly) smaller code.

All tests pass locally, including those added in pr #111.