Introduce abstraction for compact Path->MInfo maps
Similarly to how the previous commits define a FixedDir interface with both
compact and direct implementations, this commit introduces an interface for the
`FixedSingle<Path, MInfo>` structures that we use in EagerDirs (e.g. fixedOld)
since I saw that there was a ton of wasteful/redundant data in that structure.

In particular, for SKDB, each Path -> MInfo key/val pair has
 (1) the same DirName in every Path key,
 (2) a RowKey as the baseName in every Path key,
 (3) the same kinds in every such RowKey,
 (4) an MInfoSingle holding an IndexProjKey as the MInfo value,
 (5) the same kinds in every such IndexProjKey, and
 (6) exactly the same RowValues in the IndexProjKey as the Path

So, we can preserve semantics with a much more compact representation: a sorted
array of RowValues, along with a shared metadata object containing the dirName,
rowKinds, and indexKinds.  That's enough to reconstruct the Paths and MInfos on
demand.