En/decode CSV strings using JSON escaping algorithm
This deviates from 'standard' csv, if there is such a thing. We use
the json escaping algorithm. This is mostly for 2 reasons:

1. this approach is very well battle tested and accounts for more
things than csv usually considers.

2. we have logic that assumes that a row is represented on a single
line (e.g. looking for checkpoint markers, the csv parser itself!).
Standard CSV allows for line breaks in strings and this wreaks havoc.
It'll also hurt line-based buffering. This is all ok as the csv we
produce is not meant to be public facing. And ultimately we plan to
replace this with a binary format.

And it helped that we already have the json en/decoding code written
and tested.