Use ESM rather than dynamic loading for sknpm modules
Rather than dynamically loading a list of strings, produce code that
statically imports using the ESM syntax.

This allows consuming packages to use a bundler, such as vite, to
produce a single static js file. Because of ESM, this can then be
concatenated in to a single minified file (very good for http round
trip performance) and it also allows for 'tree-shaking' (discarding
e.g. functions that aren't used by the consuming code).