Get Worker working with bundler
Using this form of instantiating the worker, it becomes visible to the
bundler. This then allows the bundler to build the script the worker
loads and rewrite the file path.

Several changes were made here. Mostly to get the bundler to see the
worker load. But I also had to split out the createOnMain function as
this caused a cyclic dependency: skdb.ts -> creates worker -> loads
skdb.ts. Vite runs out of heap trying to resolve this.