Get wasm working with node and bundler
We make use of the magical import feature that vite and webpack (at
least) support: import "blah.wasm?url". This ensures the bundler sees
the file and copies it from the dependency to the dist dir. It will
rename it with a digest - which is a good thing - and this will
resolve statically to that file name.