Fix: allow skdb package to be imported without a bundler
This fixes all the tests and opens up another important way of using
the package.

The actual fix is a little sneaky/hacky - as with all things to do
with JS build/deploy :(. We use import.meta which should be available

in all contexts. vite and webpack set specific properties on this
object that we can look for. So this solution likely doesn't work for
other bundlers.