Fix: the context used by lwwEdits should be mutable
The goal of trying to keep the context readonly here was laudable but
/indexbycol/ is created lazily. By cloning and creating the dir in
the clone we end up permenantly marking /indexbycol/ as a DeletedDir.
This happens because we call update immediatley after in the w-csv hot
loop.

The overall effect is that: if replication is the first to get to
creating /indexbycol/ then it becomes permanently deleted for all
users afterward. This means that after the initial replication all
future replication is broken, and likely other things too.