Add Extension minfoType for fixed source maps
To this point, all of the specialized fixed source map types have been
_projections_, where each key maps to a value containing some subset of the
key's data.

But, sometimes we see _extensions_, where the key contains some subset of the
value's data. This commit adds logic to handle one such case, where the value
is a non-trivial extension of a singleton row key.