Don't dump system defined indices
It's not useful to see these and they'll be created when a table gets
created anyway. Having them dumped causes errors when the dump is
imported for a rebuild.

I was going to store a bit to mark the index as system created but the
namespacing is just as effective and useful in its own right.