Drop explicit stdlib dependency in SKTest
SKTest is failing to build in CircleCI here, claiming not to find the FileSystem
module.  I'm not sure why this is surfacing as a bug here, since this PR doesn't
touch anything related as far as I can tell.

But, things work without the explicit stdlib skargo dependency, so dropping it
to make the build happy.