Simplify concurrency while improving throughput
Running schedules is bounded by how fast we can vfork/clone. This
syscall is atomic blocking - it's why the python proc spends so much
time in D state. Effectively this acts as a big lock and serialises
all the schedule running (because we're non stop creating procs).

The result is that there is very little benefit in concurrently
running schedules. Instead we get parallelism by maxing out the number
of procs - and each can be forking. Unfortuantely we have to do this
at the grain of tests, but it works pretty well and simplifies the
schedule running loop.