Ack checkpoints even if they are too old to be applied
Clients are sending the data because they believe it hasn't been seen
although it has. We ack it in case they fail again, there's no need to
send once more.

This was motivated by the simulator wanting to wait on write-csv
processing data regardless of whether it is applied. And should also
be good for clients on e.g. flaky network connections.