[skc] Add support for `sklib` format.
The sklib format contains a compiled skip package. While the skip code
itself cannot be compiled in isolation, possible C extensions
can. Thus, we store arbitrary object files along with the skip source
code within a ranlib archive.
An extra `lib.skmeta` file is stored, containing various metadata
about the package, including the name of the package, its direct
dependencies, and the required linker arguments.

Dependencies to sklib packages are expressed with the `skc` `-l` flag,
which has two forms:
  - `-l foo=/path/to/some.sklib`, and
  - `-l foo`.

The latter will search for a package named `foo` within the search
path (comprised of the `sysroot`, and any path provided with the `-L`
flag). The `sysroot` (which can be overridden through the `--sysroot`
flag) defaults to `../lib/` relative to the location of the `skc`
binary.

Only direct dependencies should be specified with `-l`. This is mainly
for two reasons: 1. ergonomics, as it lifts the burden of determining
the dependency graph manually, and 2. we might want to allow multiple
versions of a given package to coexist in a dependency graph (though
this will require name-spaced packages). The exact dependencies of a
given package are stored in the `lib.skmeta`, allowing `skc` to fetch
all dependencies in a deterministic way.

In order to improve ergonomics, `-l std` is implicitly added unless
the `--no-std` flag, or an explicit dependency to `std` is
supplied. This enables building a skip binary through a simple
invocation: `skc foo.sk -o foo`, compared to the previously required
`skc foo.sk path/stdlib/src/**/*.sk -o foo`

Building sklib packages is done by specifying the `--sklib-name=foo`
flag to `skc`.
E.g. `skc --sklib-name=sqlparser -o sqlparser-dev.sklib src/**/*.sk`.
This will run the front end on the provided sources, and output an
`sklib` archive.

Due to the reactive nature of the compiler, we have to take care of
cache invalidation for source files. The simple heuristic used up to
this point consists of keeping only the analysis for the intersection
of collections of provided source files across two successive `skc`
invocations`, and re-triggering analysis for files that have
changed (checked by their `mtime`).
For instance, `skc foo.sk bar.sk baz.sk --data state.db` followed by `skc
foo.sk bar.sk foobar.sk --data state.db` would get rid of the analysis
of `bar.sk` (and that of `foo.sk` if its `mtime` has changed between
the two invocations).
This is an issue when building a whole dependency graph by
successively invoking `skc` to generate sklibs in topological order,
as analysis for sources of sklibs in independent sub-graphs would get
invalidated, causing redundant work.

Hence, we need to invalidate more diligently. The current
implementation uses the following heuristic:
- when building an sklib, only invalidate files from the *current
library*, keeping analysis around for all other libraries. This begs
the question of what is the current library? For now, we rely solely
on the value of `--sklib-name` provided on the command line. This
would need to change if we want to support compilation graphs with
several versions of a given library (which would require name-spacing
packages), or if we want to support sharing analysis between target
architectures (which is currently not possible because the front end
uses the pointer width).
- When building a binary, pretend to be building a library with an
special name. (Most binaries are not built this way, as the
corresponding library is generated first, and generating the binary
piggy-backs on the front end analysis of the library and triggering
the back end by picking an entry point. Building binaries by directly
invoking `skc foo.sk -o foo` is supported for convenience but most
projects should rely on `skargo` to handle the build.)