[soak] Make soak expectation checks more robust
The expectation checks assume that checkpoints are delivered
atomically with updates that were written alongside them.

In the case of resets (or now rebuilds) it's possible for the
checkpoint table to be rebuilt independently of the other tables. This
makes it go back in time independently of other tables. The client may
have seen and moved on from the updates that were written alongside a
checkpoint that is being delivered. This scenario breaks the
expectation checks.

This diff fixes this by keeping a history of what has been seen for
tables where rows are updated.