[soak] Make privacy checks more robust in soak
Privacy checks were failing for the same reason (see parent): the
checkpoints could go back in time.

We cannot (easily) reuse the same strategy of keeping history here -
and it's hard enough to reason about as it is - so we just run the
check if we haven't seen this watermark yet.