[soak] Don't check expectations when a table is rebuilding
Soak was failing because it would receive (on a reconnect) changes for
several tables including at least one that was a rebuild and one that
was a new checkpoint. It would try to check the expectations against a
table that was wiped out.

This isn't fair, and is inherent to the approach of rebuilding tables,
so we check for this situation now.