Add safe guard around rebuild processing for the server
The whole point of moving from push based 'resets' to pull based
'rebuilds' is to protect the server. We do not want to wipe out huge
amounts of data in a shared environment which is being
accessed/modified concurrently. Although we _think_ the logic is good,
it's dangerous. Instead we push this risk to each client.

This change makes processing rebuilds opt-in, which of course we only
do in the client.