[sql] Add support for omitting columns in INSERT SELECT
Values for columns that are either PRIMARY KEYs, NULLABLE, or have a
DEFAULT may be omitted when inserting rows. This PR adds support for
such omissions when INSERTing the rows returned by a SELECT.

The `computeInsert` function is refactored to expose
`computeInsertRow`, which contains the functionality to insert omitted
defaults, etc. This `computeInsertRow` function is then also called
when evaluating INSERT SELECT, in the mapper over the rows returned by
the query that inserts them into the table. Also, the error checking
code is adjusted accordingly.

closes #165