Create groups in a single transaction
I want group creation to be atomic. This prevents interleaving
rebuilds from corrupting state. In particular it's possible for a
rebuild to interleave and the client hold on to some updates that
haven't yet been ack'd. The server sends back down previous versions
of rows and this breaks uniqueness constraints - causing the group
creator to get in to a permanent bad state.