Allow atomic group creation in the privacy rules
For atomic group creation to be possible we need to break the cycle:
the group insert checks that the user has permission to write as this
author (which is the group and there are no permissions for it yet)
and the permission write fails because the group does not yet exist.
The group insert also checks the group exists which it can't because
it's itself.

This diff allows:

- a permission to be created for a group that does not yet exist.
- a group to be created with itself as the author