Check privacy rules in w-csv as they are updated
This is the last piece to unlocking atomic group creation. We need
write-csv to apply the updates in order and using rules that it might
just have updated (instead of a snapshot before any updates are
written). This has the side effect of bringing all privacy checks
in to the lock. This is likely very good for correctness (no weird
races or concurrency issues) but might be bad for perf.